Fix edit mode bugs
Implement error handling for incorrectly formatted field values in edit mode.

Fix bug where mob stats were defaulting to base values when edited.

Fix bug where detailed panel view would crash when editing item entities.

+++ src/main/java/seedu/address/experimental/storage/JsonAdaptedMob.java
@@ -17,18 +17,19 @@ import seedu.address.model.tag.Tag;
 
 /***/
 public class JsonAdaptedMob {
+
     private final String name;
     private final JsonAdaptedStats stats;
-    private final int challengeRating;
+    private final float challengeRating;
     private final boolean isLegendary;
     private final List<JsonAdaptedTag> tagged = new ArrayList<>();
 
     /***/
     @JsonCreator
     JsonAdaptedMob(@JsonProperty("name") String name, @JsonProperty("stats") JsonAdaptedStats stats,
-                   @JsonProperty("challengeRating") int challengeRating,
-                   @JsonProperty("isLegendary") boolean isLegendary,
-                   @JsonProperty("tagged") List<JsonAdaptedTag> tagged) {
+            @JsonProperty("challengeRating") float challengeRating,
+            @JsonProperty("isLegendary") boolean isLegendary,
+            @JsonProperty("tagged") List<JsonAdaptedTag> tagged) {
         this.name = name;
         this.stats = stats;
         this.challengeRating = challengeRating;
@@ -37,6 +38,7 @@ public class JsonAdaptedMob {
             this.tagged.addAll(tagged);
         }
     }
+
     /***/
     public JsonAdaptedMob(Mob source) {
         name = source.getName().fullName;
@@ -47,6 +49,7 @@ public class JsonAdaptedMob {
                 .map(JsonAdaptedTag::new)
                 .collect(Collectors.toList()));
     }
+
     /***/
     public Mob toModelType() throws IllegalValueException {
         // dont care
